Split pulse property computation out of hitfinder
