+++ pax/config/XENON100.ini
@@ -15,6 +15,7 @@ dsp = [
             # Do some sanity checks / cleaning on pulses
             'CheckPulses.SortPulses',
             'CheckPulses.ConcatenateAdjacentPulses',
+            'PulseProperties.PulseProperties',
             'CheckPulses.CheckBoundsAndCount',
 
             # 'FakeTrigger.FakeTrigger',
@@ -69,12 +70,13 @@ pre_analysis = [
                ]
 
 
+[PulseProperties]
+# Compute baseline on first n and last n samples in pulse.
+baseline_samples = 50
+
 
 [HitFinder]
 # For detailed description of what these settings do, see the documentation / plugin docstring.
-
-# Compute baseline on first n samples in pulse:
-initial_baseline_samples = 40
 
 # Max hits to look for in each pulse: rest will be ignored
 max_hits_per_pulse = 500
