+++ pax/config/XENON100_LED.ini
@@ -2,7 +2,8 @@
 parent_configuration = 'XENON100'
 
 # Stop after the hitfinder - everything else is meaningless/counterproductive
-dsp = ['CheckPulses.CheckBounds',
+dsp = ['PulseProperties.PulseProperties',
+       'CheckPulses.CheckBoundsAndCount',
        'HitFinder.FindHits',
       ]
 compute_properties = []
