+++ pax/config/XENON1T.ini
@@ -12,9 +12,12 @@ decoder_plugin = 'Pickle.DecodeZPickle'
 # Digital signal processing - if you want to repeat any of this,
 # you need to go back to raw data and start from scratch
 dsp = [
-            # Do some sanity checks / cleaning on pulses
+            # Compute the basic pulse properties
             'CheckPulses.SortPulses',
             'CheckPulses.ConcatenateAdjacentPulses',
+            'PulseProperties.PulseProperties',
+
+            # Truncate pulses outside the event and count them
             'CheckPulses.CheckBoundsAndCount',
 
             # Find individual hits
@@ -101,11 +104,13 @@ waveforms_to_plot = (
     )
 
 
+[PulseProperties]
+# Compute baseline on first n and last n samples in pulse.
+baseline_samples = 40
+
+
 [HitFinder]
 # For detailed description of what these settings do, see the documentation / plugin docstring.
-
-# Compute baseline on first n samples in pulse:
-initial_baseline_samples = 40
 
 # Max hits to look for in each pulse: rest will be ignored
 max_hits_per_pulse = 500
