+++ pax/config/XENON1T_LED.ini
@@ -3,6 +3,7 @@ parent_configuration = 'XENON1T'
 
 # Stop after the hitfinder - everything else is meaningless/counterproductive
 dsp = [#'CheckPulses.CheckBounds',   # Temporarily disabled, see #315
+       'PulseProperties.PulseProperties',
        'HitFinder.FindHits',
       ]
 compute_properties = []
