+++ pax/config/XENON1T_MV.ini
@@ -13,7 +13,8 @@ dsp = [
             # Do some sanity checks / cleaning on pulses
             'CheckPulses.SortPulses',
             'CheckPulses.ConcatenateAdjacentPulses',
-            'CheckPulses.CheckBounds',
+            'PulseProperties.PulseProperties',
+            'CheckPulses.CheckBoundsAndCount',
 
             # Find individual hits
             'HitFinder.FindHits',
