+++ pax/config/reduce_raw_data.ini
@@ -2,7 +2,7 @@
 ; the --event_numbers_file switch to reduce raw data files
 
 [pax]
-;  don't process, just reduce data
+;  don't process, just change the data format
 dsp = []
 transform = []
 compute_properties = []
