+++ tests/test_hitfinder.py
@@ -6,6 +6,7 @@ import os
 import sys
 sys.path.append(os.path.join(utils.PAX_DIR, 'plugins', 'signal_processing'))
 import HitFinder  # flake8: noqa
+import PulseProperties   # flake8: noqa
 
 
 class TestHitFinder(unittest.TestCase):
@@ -23,8 +24,9 @@ class TestHitFinder(unittest.TestCase):
                                       'pax': {
                                           'plugin_group_names': ['test'],
                                           'encoder_plugin': None,
-                                          'test':               'HitFinder.FindHits'}})
-        plugin = self.pax.get_plugin_by_name('FindHits')
+                                          'decoder_plugin': None,
+                                          'test':               ['PulseProperties.PulseProperties',
+                                                                 'HitFinder.FindHits']}})
         for test_w, hit_bounds, pulse_min, pulse_max in (
             # Keep in mind the hitfinder flips the pulse...
             [np.zeros(100), [], 0, 0],
@@ -36,14 +38,14 @@ class TestHitFinder(unittest.TestCase):
             [self.peak_at(70, amplitude=-100, width=4) + self.peak_at(80, amplitude=-100, width=4),
              [[70, 73], [80, 83]], 0, 100],
         ):
-            e = datastructure.Event(n_channels=plugin.config['n_channels'],
+            e = datastructure.Event(n_channels=self.pax.config['DEFAULT']['n_channels'],
                                     start_time=0,
-                                    sample_duration=plugin.config['sample_duration'],
+                                    sample_duration=self.pax.config['DEFAULT']['sample_duration'],
                                     stop_time=int(1e6),
                                     pulses=[dict(left=0,
                                                  raw_data=np.array(test_w).astype(np.int16),
                                                  channel=1)])
-            e = plugin.transform_event(e)
+            e = self.pax.process_event(e)
             self.assertEqual(hit_bounds, [[hit['left'], hit['right']] for hit in e.all_hits])
             self.assertEqual(pulse_min, e.pulses[0].minimum)
             self.assertEqual(pulse_max, e.pulses[0].maximum)
@@ -84,7 +86,8 @@ class TestHitFinder(unittest.TestCase):
             [42],
         ):
             w = np.array(w, dtype=np.float64)
-            baseline, baseline_increase, noise, min_w, max_w = HitFinder.compute_pulse_properties(w, baseline_samples=10)
+            results = PulseProperties.compute_pulse_properties(w, baseline_samples=10)
+            baseline, baseline_increase, noise, min_w, max_w = results
             bl_before = np.mean(w[:min(len(w), 10)])
             bl_after = np.mean(w[-min(len(w), 10):])
             self.assertEqual(baseline, max(bl_before, bl_after))
